+++ src/app/accommodations/bookings/booking/special-request.service.ts
@@ -1,14 +1,17 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, switchMap, take } from 'rxjs/operators';
 import { SpecialRequest } from './special-request.model';
+import { AuthService } from 'src/app/auth/auth.service';
 
 @Injectable({
     providedIn: 'root'
 })
 export class SpecialRequestService {
-    constructor(private http: HttpClient) { }
+    private baseUrl: string = 'https://accommodation-app-a89f8-default-rtdb.europe-west1.firebasedatabase.app/special_requests';
+
+    constructor(private http: HttpClient, private authService: AuthService) { }
 
     addSpecialRequest(bookingId: string, description: string): Observable<{ name: string }> {
         const newRequest: SpecialRequest = {
@@ -16,32 +19,56 @@ export class SpecialRequestService {
             booking_id: bookingId,
             description: description
         };
-        return this.http.post<{ name: string }>(
-            `https://accommodation-app-a89f8-default-rtdb.europe-west1.firebasedatabase.app/special_requests.json`,
-            { ...newRequest, id: null }
+
+        return this.authService.user.pipe(
+            take(1),
+            switchMap(user => {
+                if (!user || !user.token) {
+                    throw new Error('No user found or user is not authenticated!');
+                }
+                return this.http.post<{ name: string }>(
+                    `${this.baseUrl}.json?auth=${user.token}`,
+                    { ...newRequest, id: null }
+                );
+            })
         );
     }
 
     getSpecialRequestsByBookingId(bookingId: string): Observable<SpecialRequest[]> {
-        return this.http.get<{ [key: string]: SpecialRequest }>(
-            `https://accommodation-app-a89f8-default-rtdb.europe-west1.firebasedatabase.app/special_requests.json?orderBy="booking_id"&equalTo="${bookingId}"`
-        ).pipe(
-            map((specialRequestData: { [key: string]: SpecialRequest }) => {
-                const specialRequests: SpecialRequest[] = [];
-                for (const key in specialRequestData) {
-                    if (specialRequestData.hasOwnProperty(key)) {
-                        specialRequests.push({ ...specialRequestData[key], id: key });
-                    }
+        return this.authService.user.pipe(
+            take(1),
+            switchMap(user => {
+                if (!user || !user.token) {
+                    throw new Error('No user found or user is not authenticated!');
                 }
-                return specialRequests;
+                return this.http.get<{ [key: string]: SpecialRequest }>(
+                    `${this.baseUrl}.json?orderBy="booking_id"&equalTo="${bookingId}"&auth=${user.token}`
+                ).pipe(
+                    map((specialRequestData: { [key: string]: SpecialRequest }) => {
+                        const specialRequests: SpecialRequest[] = [];
+                        for (const key in specialRequestData) {
+                            if (specialRequestData.hasOwnProperty(key)) {
+                                specialRequests.push({ ...specialRequestData[key], id: key });
+                            }
+                        }
+                        return specialRequests;
+                    })
+                );
             })
         );
     }
 
     deleteSpecialRequest(requestId: string): Observable<void> {
-        return this.http.delete<void>(
-            `https://accommodation-app-a89f8-default-rtdb.europe-west1.firebasedatabase.app/special_requests/${requestId}.json`
+        return this.authService.user.pipe(
+            take(1),
+            switchMap(user => {
+                if (!user || !user.token) {
+                    throw new Error('No user found or user is not authenticated!');
+                }
+                return this.http.delete<void>(
+                    `${this.baseUrl}/${requestId}.json?auth=${user.token}`
+                );
+            })
         );
     }
-
 }
